update to new configuration structure
